fix double

+++ Firmware/aire_apollo/src/pwmValve.h
@@ -11,22 +11,16 @@ public:
 
   pwmValve(uint8_t pin);
   virtual bool    begin();
-  virtual void    open(uint8_t percent = 100);
+  virtual void    open(double percent = 100);
   virtual void    close();
-  virtual uint8_t status() {return _percent > 0;}
+  virtual double status() {return _percent > 0;}
 
 protected:
 
-  uint8_t _percent;
+  double _percent;
   uint8_t _pin;
 
 };
-
-
-
-
-
-
 
 
 #endif
